webapi/tct-webaudio-w3c-tests/webaudio/offlineaudiocontext_suspend_and_resume.html

<!DOCTYPE html>
<!--
Copyright (c) 2016 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>WebAudio Test: OfflineAudioContext.suspend() and resume()</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Chunyan Wang" href="mailto:chunyanx.wang@intel.com">
<link rel="help" href="http://webaudio.github.io/web-audio-api/#idl-def-OfflineAudioContext">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function(t) {
  var sampleRate = 44100;
  var duration = 1;
  var offlineAudioContext = new OfflineAudioContext(1, sampleRate * duration, sampleRate);
  offlineAudioContext.suspend(-0.1)
    .then(t.step_func(function() {
      assert_unreached("unreached here when suspend time is negative");
    }))
    .catch(t.step_func(function(ex) {
      assert_equals(offlineAudioContext.state, "suspended");
      assert_equals(ex.name, "InvalidStateError");
      t.done()
    }));
}, "Check that negative suspend time is not allowed");

async_test(function(t) {
  var sampleRate = 44100;
  var duration = 1;
  var offlineAudioContext = new OfflineAudioContext(1, sampleRate * duration, sampleRate);
  offlineAudioContext.suspend(duration + 0.01)
    .then(t.step_func(function() {
      assert_unreached("unreached here when suspend time is greater than total render duration");
    }))
    .catch(t.step_func(function(ex) {
      assert_equals(ex.name, "InvalidStateError");
      t.done()
    }));
}, "Check that suspend() should throw InvalidStateError exception " +
   "when suspend time is greater than total render duration");

async_test(function(t) {
  var sampleRate = 44100;
  var duration = 1;
  var offlineAudioContext = new OfflineAudioContext(1, sampleRate * duration, sampleRate);
  offlineAudioContext.resume()
    .then(t.step_func(function() {
      assert_unreached("unreached here when the offline context has not started");
    }))
    .catch(t.step_func(function(ex) {
      assert_equals(offlineAudioContext.state, "suspended");
      assert_equals(ex.name, "InvalidStateError");
      t.done();
    }));
}, "Check that it cann't resume an offline context that has not started");

async_test(function(t) {
  var sampleRate = 44100;
  var duration = 1;
  var offlineAudioContext = new OfflineAudioContext(1, sampleRate * duration, sampleRate);

  offlineAudioContext.suspend(0.001)
    .then(t.step_func(function() {
      assert_equals(offlineAudioContext.state, "suspended");
      offlineAudioContext.resume();
    }))
    .then(t.step_func(function() {
      assert_equals(offlineAudioContext.state, "running");
    }))
    .catch(t.step_func(function(ex) {
      assert_unreached("unreached here");
    }));

  offlineAudioContext.startRendering()
    .then(t.step_func(function() {
      assert_equals(offlineAudioContext.state, "closed");
      t.done();
    }))
}, "Check that an offline context can be suspended and resumed");
</script>
